+++ Tactical/LaserOnFire/changelog.md
@@ -1,5 +1,10 @@
 Changelog of Laser on Fire, a Phoenix Point Mod by Sheepy
 
+# Version 1, 2021-02-01
+
+* Update to Scripting Library 2, which uses JavaScript instead of C#.
+* Remove Destiny III first person view and camera shake, since they have been implemented in vanilla.
+
 # Version Alpha 0.0.2020.0605
 
 * Update to new Scripting Library.